+++ backend/utils/github_test.go
@@ -1,6 +1,7 @@
 package utils
 
 import (
+	"github.com/diggerhq/digger/libs/git_utils"
 	"os"
 	"testing"
 
@@ -9,14 +10,14 @@ import (
 
 func TestGithubCloneWithInvalidTokenThrowsErr(t *testing.T) {
 	f := func(d string) error { return nil }
-	err := CloneGitRepoAndDoAction("https://github.com/diggerhq/private-repo", "main", "", "invalid-token", "", f)
+	err := git_utils.CloneGitRepoAndDoAction("https://github.com/diggerhq/private-repo", "main", "", "invalid-token", "", f)
 	assert.NotNil(t, err)
 }
 
 func TestGithubCloneWithPublicRepoThrowsNoError(t *testing.T) {
 	token := os.Getenv("GITHUB_PAT_TOKEN")
 	f := func(d string) error { return nil }
-	err := CloneGitRepoAndDoAction("https://github.com/diggerhq/digger", "develop", "", token, "", f)
+	err := git_utils.CloneGitRepoAndDoAction("https://github.com/diggerhq/digger", "develop", "", token, "", f)
 	assert.Nil(t, err)
 }
 
@@ -27,13 +28,13 @@ func TestGithubCloneWithPrivateRepoAndValidTokenThrowsNoError(t *testing.T) {
 		return
 	}
 	f := func(d string) error { return nil }
-	err := CloneGitRepoAndDoAction("https://github.com/diggerhq/infra-gcp", "main", "", token, "", f)
+	err := git_utils.CloneGitRepoAndDoAction("https://github.com/diggerhq/infra-gcp", "main", "", token, "", f)
 	assert.Nil(t, err)
 }
 
 func TestGithubCloneWithInvalidBranchThrowsError(t *testing.T) {
 	token := os.Getenv("GITHUB_PAT_TOKEN")
 	f := func(d string) error { return nil }
-	err := CloneGitRepoAndDoAction("https://github.com/diggerhq/digger", "not-a-branch", "", token, "", f)
+	err := git_utils.CloneGitRepoAndDoAction("https://github.com/diggerhq/digger", "not-a-branch", "", token, "", f)
 	assert.NotNil(t, err)
 }
