+++ background/projects-refresh-service/.gitignore
@@ -1,0 +1,1 @@
+main
