+++ background/projects-refresh-service/go.mod
@@ -1,0 +1,3 @@
+module github.com/diggerhq/digger/backgorund/projects-refresh-service
+
+go 1.24.0
