+++ ee/drift/utils/github.go
@@ -5,10 +5,10 @@ import (
 	"encoding/base64"
 	"fmt"
 	"github.com/bradleyfalzon/ghinstallation/v2"
-	utils2 "github.com/diggerhq/digger/backend/utils"
 	"github.com/diggerhq/digger/ee/drift/dbmodels"
 	github2 "github.com/diggerhq/digger/libs/ci/github"
 	dg_configuration "github.com/diggerhq/digger/libs/digger_config"
+	utils2 "github.com/diggerhq/digger/libs/git_utils"
 	"github.com/diggerhq/digger/next/utils"
 	"github.com/dominikbraun/graph"
 	"github.com/google/go-github/v61/github"
