+++ go.work
@@ -2,6 +2,7 @@ go 1.24.0
 
 use (
 	./backend
+	./background/projects-refresh-service
 	./cli
 	./cli_e2e
 	./dgctl
