+++ next/services/config.go
@@ -2,8 +2,8 @@ package services
 
 import (
 	"fmt"
-	utils3 "github.com/diggerhq/digger/backend/utils"
 	dg_configuration "github.com/diggerhq/digger/libs/digger_config"
+	utils3 "github.com/diggerhq/digger/libs/git_utils"
 	"github.com/diggerhq/digger/next/dbmodels"
 	"github.com/diggerhq/digger/next/utils"
 	"log"
